+++ src/main/java/com/transferwise/acorn/models/BalancesAfter.java
@@ -1,7 +1,9 @@
 package com.transferwise.acorn.models;
 
+import java.math.BigDecimal;
+
 public class BalancesAfter{
     public int id;
-    public double value;
+    public BigDecimal value;
     public String currency;
 }
